# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/login": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "로그인 페이지 렌더링",
        "description": "로그인 폼이 있는 HTML 페이지를 렌더링합니다.",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "로그인 페이지 렌더링 성공"
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "사용자 로그인",
        "description": "사용자의 이메일과 비밀번호를 받아 로그인 처리 후 세션에 사용자 정보를 저장합니다.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "로그인 정보",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "로그인 성공. 대시보드로 리다이렉트됩니다."
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "인증 실패. 이메일 또는 비밀번호가 올바르지 않습니다."
          }
        }
      }
    },
...
    "/auth/register": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "회원가입 페이지 렌더링",
        "description": "회원가입 폼이 있는 HTML 페이지를 렌더링합니다.",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "회원가입 페이지 렌더링 성공"
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "신규 사용자 회원가입",
        "description": "새로운 사용자를 시스템에 등록합니다.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "회원가입 정보",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "홍길동"
                },
                "email": {
                  "type": "string",
                  "example": "newuser@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                },
                "role": {
                  "type": "string",
                  "example": "STUDENT"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "회원가입 성공"
          },
          "400": {
            "description": "잘못된 요청 (예: 필수 필드 누락)"
          },
          "409": {
            "description": "이미 존재하는 이메일"
          }
        }
      }
    },
...
    "/me/dashboard/": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "역할별 대시보드 조회",
        "description": "로그인한 사용자의 역할(학생/강사/관리자)에 맞는 대시보드 페이지를 렌더링합니다.",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "accept",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "대시보드 렌더링 성공"
          },
          "401": {
            "description": "로그인 필요"
          }
        }
      }
    },
...
  }
}